Further Double -> Float and x, y to Point

+++ korge-sandbox/src/commonMain/kotlin/samples/MainText.kt
@@ -1,20 +1,20 @@
 package samples
 
-import korlibs.time.*
-import korlibs.korge.scene.*
-import korlibs.korge.view.*
 import korlibs.image.bitmap.effect.*
 import korlibs.image.color.*
 import korlibs.image.font.*
 import korlibs.image.paint.*
 import korlibs.image.text.*
 import korlibs.io.lang.*
+import korlibs.korge.scene.*
+import korlibs.korge.view.*
 import korlibs.math.geom.*
+import korlibs.time.*
 
 class MainText : ScaledScene(512, 160) {
     override suspend fun SContainer.sceneMain() {
         val font = BitmapFont(
-            DefaultTtfFont, 64.0,
+            DefaultTtfFont, 64f,
             paint = LinearGradientPaint(0, 0, 0, 50).add(0.0, Colors.CADETBLUE).add(1.0, Colors.PURPLE),
             effect = BitmapEffect(
                 dropShadowX = 2,
@@ -27,7 +27,7 @@ class MainText : ScaledScene(512, 160) {
         var offset = 0.degrees
         var version = 0
         //text("Hello World!", font = font, textSize = 64.0, alignment = TextAlignment.BASELINE_LEFT, renderer = CreateStringTextRenderer({ version++ }) { reader: WStringReader, c: Int, g: GlyphMetrics, advance: Double ->
-        val text = text("Hello World!", font = font, textSize = 64.0, alignment = TextAlignment.BASELINE_LEFT, renderer = CreateStringTextRenderer({ version++ }) { reader: WStringReader, c: Int, g: GlyphMetrics, advance: Double ->
+        val text = text("Hello World!", font = font, textSize = 64f, alignment = TextAlignment.BASELINE_LEFT, renderer = CreateStringTextRenderer({ version++ }) { reader: WStringReader, c: Int, g: GlyphMetrics, advance: Float ->
             val sin = sind(offset + (reader.position * 360 / reader.length).degrees)
             transform = Matrix()
                 .rotated(15.degrees)
